+++ packages/bc_ur_dart/lib/src/models/btc/gspl_signature.dart
@@ -1,8 +1,10 @@
 import 'dart:typed_data';
 
 import 'package:bc_ur_dart/src/models/btc/gspl_tx_data.dart';
+import 'package:bc_ur_dart/src/models/btc/gspl_sign_request.dart';
 import 'package:cbor/cbor.dart';
 import 'package:bc_ur_dart/src/ur.dart';
+import 'package:crypto_wallet_util/transaction.dart' show GsplTxData;
 
 const String BTC_SIGNATURE = 'BTC-SIGNATURE';
 
@@ -18,8 +20,19 @@ class GsplSignatureUR extends UR {
     final data = ur.decodeCBOR() as CborMap;
 
     final uuid = Uint8List.fromList((data[CborSmallInt(1)] as CborBytes).bytes);
-    final gspl = GsplTxData.fromCbor(data: data[CborSmallInt(2)] as CborMap);
+    final gspl = getGsplTxDataFromCbor(data: data[CborSmallInt(2)] as CborMap);
 
     return GsplSignatureUR(uuid: uuid, gspl: gspl, type: ur.type, payload: ur.payload);
+  }
+
+  factory GsplSignatureUR.fromSignature({required GsplSignRequestUR request, required GsplTxData gspl}) {
+    final ur = UR.fromCBOR(
+      type: BTC_SIGNATURE,
+      value: CborMap({
+        CborSmallInt(1): CborBytes(request.uuid, tags: [37]),
+        CborSmallInt(2): gspl.toCbor(),
+      })
+    );
+    return GsplSignatureUR(uuid: request.uuid, gspl: gspl, type: ur.type, payload: ur.payload);
   }
 }