+++ packages/bc_ur_dart/CHANGELOG.md
@@ -61,3 +61,12 @@
 ## 0.1.16
 
 - Fix: Fixed CryptoMultiAccountsUR XRP get from wallet.
+
+## 0.1.17
+
+- Fix: Fixed address and signHashType parse error of GsplItem.
+- Fix: Refactor some code about gspl
+
+## 0.1.18
+
+- Update: Update crypto_wallet_util pubspec version.