+++ packages/crypto_wallet_util/CHANGELOG.md
@@ -164,3 +164,16 @@
 ### Update
 
 - fix psbt signer error.
+
+## [1.1.8] - 2025-07-28
+### Update
+
+- Add gspl signer.
+- Add doge wallet
+- Add ltc wallet
+- Add bch wallet
+
+## [1.1.9] - 2025-07-28
+## Update
+
+- Update gitignore file.